+++ pgpm/env/src/index.ts
@@ -11,6 +11,6 @@ export {
 } from './config';
 export type { WorkspaceType } from './config';
 export { getEnvVars, getNodeEnv, parseEnvBoolean, parseEnvNumber } from './env';
-export { walkUp, mergeArraysUnique } from './utils';
+export { walkUp, replaceArrays } from './utils';
 
 export type { PgpmOptions, PgTestConnectionOptions, DeploymentOptions } from '@pgpmjs/types';